kairos-server: add the sse event trigger
