+++ kairos-server/src/lib.rs
@@ -1,21 +1,20 @@
 pub mod config;
 pub mod errors;
-pub mod on_deploy;
 pub mod routes;
 pub mod state;
 
 mod l1_sync;
 mod utils;
 
-use backoff::{future::retry, Error, ExponentialBackoff};
+use std::collections::HashSet;
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::mpsc;
+use tokio::sync::RwLock;
 
+use casper_client::types::DeployHash;
+use casper_client_hashing::Digest;
 use casper_deploy_notifier::DeployNotifier;
-use casper_event_toolkit::fetcher::Fetcher;
-use casper_event_toolkit::metadata::CesMetadataRef;
-use casper_event_toolkit::rpc::client::CasperClient;
 use casper_types::ContractHash;
 
 use axum::Router;
@@ -50,7 +49,7 @@ pub fn app_router(state: ServerState) -> Router {
         .with_state(state)
 }
 
-pub async fn run_l1_sync(server_state: Arc<ServerStateInner>) {
+pub async fn run_l1_sync(server_state: ServerState) {
     // Extra check: make sure the default dummy value of contract hash was changed.
     let contract_hash = server_state.server_config.kairos_demo_contract_hash;
     if contract_hash == ContractHash::default() {
@@ -61,36 +60,23 @@ pub async fn run_l1_sync(server_state: Arc<ServerStateInner>) {
     }
 
     // Initialize L1 synchronizer.
-    let l1_sync_service = L1SyncService::new(server_state).await.unwrap_or_else(|e| {
-        panic!("Event manager failed to initialize: {}", e);
-    });
+    let l1_sync_service = Arc::new(
+        L1SyncService::new(server_state.clone())
+            .await
+            .unwrap_or_else(|e| {
+                panic!("Event manager failed to initialize: {}", e);
+            }),
+    );
 
     // Run periodic synchronization.
-    // TODO: Add additional SSE trigger.
+    let l1_sync_service_clone = l1_sync_service.clone();
     tokio::spawn(async move {
-        l1_sync::interval_trigger::run(l1_sync_service.into()).await;
+        l1_sync::interval_trigger::run(l1_sync_service_clone).await;
     });
-}
-
-pub async fn run(config: ServerConfig) {
-    let listener = tokio::net::TcpListener::bind(config.socket_addr)
-        .await
-        .unwrap_or_else(|err| panic!("Failed to bind to address {}: {}", config.socket_addr, err));
-    tracing::info!("listening on `{}`", listener.local_addr().unwrap());
-
-    let state = Arc::new(ServerStateInner {
-        batch_state_manager: BatchStateManager::new_empty(),
-        server_config: config.clone(),
-    });
-
-    run_l1_sync(state.clone()).await;
-
-    let app = app_router(state.clone());
 
     // deploy notifier
     let (tx, mut rx) = mpsc::channel(100);
-    let mut deploy_notifier = DeployNotifier::new(config.casper_sse.as_str());
-
+    let mut deploy_notifier = DeployNotifier::new(server_state.server_config.casper_sse.as_str());
     tokio::spawn(async move {
         loop {
             if let Err(e) = deploy_notifier.connect().await {
@@ -107,45 +93,29 @@ pub async fn run(config: ServerConfig) {
             tokio::time::sleep(Duration::from_secs(5)).await;
         }
     });
-
     // deploy listener/ callback
     tokio::spawn(async move {
-        let casper_client = CasperClient::new(state.server_config.casper_rpc.as_str());
-        let metadata = retry(ExponentialBackoff::default(), || async {
-            CesMetadataRef::fetch_metadata(
-                &casper_client,
-                &state
-                    .server_config
-                    .kairos_demo_contract_hash
-                    .to_formatted_string(),
-            )
-            .await
-            .map_err(Error::transient)
-        })
-        .await
-        .expect("Failed to fetch the demo contracts event metadata");
-
-        let fetcher = Fetcher {
-            client: CasperClient::default_mainnet(),
-            ces_metadata: metadata,
-        };
-
-        let schemas = retry(ExponentialBackoff::default(), || async {
-            fetcher.fetch_schema().await.map_err(Error::transient)
-        })
-        .await
-        .expect("Failed to fetch the demo contracts event schema");
-
-        while let Some(notification) = rx.recv().await {
-            on_deploy::on_deploy_notification(
-                &fetcher,
-                &schemas,
-                Arc::clone(&state),
-                &notification,
-            )
-            .await;
+        while let Some(_notification) = rx.recv().await {
+            l1_sync::interval_trigger::run(l1_sync_service.clone()).await;
         }
     });
+}
+
+pub async fn run(config: ServerConfig) {
+    let listener = tokio::net::TcpListener::bind(config.socket_addr)
+        .await
+        .unwrap_or_else(|err| panic!("Failed to bind to address {}: {}", config.socket_addr, err));
+    tracing::info!("listening on `{}`", listener.local_addr().unwrap());
+
+    let state = Arc::new(ServerStateInner {
+        batch_state_manager: BatchStateManager::new_empty(),
+        server_config: config.clone(),
+        known_deposit_deploys: RwLock::new(HashSet::new()),
+    });
+
+    run_l1_sync(state.clone()).await;
+
+    let app = app_router(state.clone());
 
     axum::serve(listener, app)
         .with_graceful_shutdown(shutdown_signal())
